+++ cmd/janus/server.go
@@ -25,6 +25,10 @@ import (
 	_ "github.com/hellofresh/janus/pkg/plugin/requesttransformer"
 	_ "github.com/hellofresh/janus/pkg/plugin/responsetransformer"
 
+	// dynamically registered auth providers
+	_ "github.com/hellofresh/janus/pkg/jwt/basic"
+	_ "github.com/hellofresh/janus/pkg/jwt/github"
+
 	// internal plugins
 	_ "github.com/hellofresh/janus/pkg/loader"
 	_ "github.com/hellofresh/janus/pkg/web"