+++ pkg/jwt/handler.go
@@ -1,12 +1,12 @@
 package jwt
 
 import (
-	"encoding/json"
-	"errors"
 	"net/http"
 	"time"
 
 	"github.com/dgrijalva/jwt-go"
+	"github.com/hellofresh/janus/pkg/config"
+	"github.com/hellofresh/janus/pkg/jwt/provider"
 	"github.com/hellofresh/janus/pkg/render"
 )
 
@@ -24,24 +24,20 @@ type Login struct {
 // Login can be used by clients to get a jwt token.
 // Payload needs to be json in the form of {"username": "<USERNAME>", "password": "<PASSWORD>"}.
 // Reply will be of the form {"token": "<TOKEN>"}.
-func (j *Handler) Login() http.HandlerFunc {
+func (j *Handler) Login(config config.Credentials) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
-		var loginValues Login
+		factory := provider.Factory{}
+		p := factory.Build(r.URL.Query().Get("provider"), config)
 
-		if json.NewDecoder(r.Body).Decode(&loginValues) != nil {
-			j.Guard.Unauthorized(w, r, errors.New("missing username or password"))
+		verified, err := p.Verify(r)
+		if err != nil {
+			render.JSON(w, http.StatusBadRequest, err.Error())
 			return
 		}
 
-		userID, ok := j.Guard.Authenticator(loginValues.Username, loginValues.Password)
-
-		if !ok {
-			j.Guard.Unauthorized(w, r, errors.New("invalid username or password"))
+		if !verified {
+			render.JSON(w, http.StatusUnauthorized, "verification failed")
 			return
-		}
-
-		if userID == "" {
-			userID = loginValues.Username
 		}
 
 		if 0 == j.Guard.Timeout {
@@ -49,11 +45,9 @@ func (j *Handler) Login() http.HandlerFunc {
 		}
 
 		expire := time.Now().Add(j.Guard.Timeout)
-
-		tokenString, err := IssueAdminToken(j.Guard.SigningMethod, userID, j.Guard.Timeout)
-
+		tokenString, err := IssueAdminToken(j.Guard.SigningMethod, map[string]interface{}{}, j.Guard.Timeout)
 		if err != nil {
-			j.Guard.Unauthorized(w, r, errors.New("problem issuing JWT"))
+			render.JSON(w, http.StatusUnauthorized, "problem issuing JWT")
 			return
 		}
 
@@ -75,7 +69,7 @@ func (j *Handler) Refresh() http.HandlerFunc {
 		origIat := int64(claims["iat"].(float64))
 
 		if origIat < time.Now().Add(-j.Guard.MaxRefresh).Unix() {
-			j.Guard.Unauthorized(w, r, errors.New("token is expired"))
+			render.JSON(w, http.StatusUnauthorized, "token is expired")
 			return
 		}
 
@@ -95,7 +89,7 @@ func (j *Handler) Refresh() http.HandlerFunc {
 		// currently only HSXXX algorithms are supported for issuing admin token, so we cast key to bytes array
 		tokenString, err := newToken.SignedString([]byte(j.Guard.SigningMethod.Key))
 		if err != nil {
-			j.Guard.Unauthorized(w, r, errors.New("create JWT Token failed"))
+			render.JSON(w, http.StatusUnauthorized, "create JWT Token failed")
 			return
 		}