+++ pkg/jwt/middleware.go
@@ -1,18 +1,20 @@
 package jwt
 
 import (
-	"context"
-	"errors"
 	"net/http"
 
-	jwt "github.com/dgrijalva/jwt-go"
+	"github.com/hellofresh/janus/pkg/render"
+	log "github.com/sirupsen/logrus"
 )
 
 // Payload Represents the context key
 type Payload struct{}
 
-// UserID Represents the user context key
-type UserID struct{}
+// User represents a logged in user
+type User struct {
+	Username string
+	Email    string
+}
 
 // Middleware struct contains data and logic required for middleware functionality
 type Middleware struct {
@@ -28,21 +30,10 @@ func NewMiddleware(config Guard) *Middleware {
 func (m *Middleware) Handler(handler http.Handler) http.Handler {
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		parser := Parser{m.Guard.ParserConfig}
-		token, err := parser.ParseFromRequest(r)
-
+		_, err := parser.ParseFromRequest(r)
 		if err != nil {
-			m.Guard.Unauthorized(w, r, err)
-			return
-		}
-
-		claims := token.Claims.(jwt.MapClaims)
-
-		id := claims["id"].(string)
-		context.WithValue(r.Context(), Payload{}, claims)
-		context.WithValue(r.Context(), UserID{}, id)
-
-		if !m.Guard.Authorizator(id, w, r) {
-			m.Guard.Unauthorized(w, r, errors.New("you don't have permission to access"))
+			log.WithError(err).Debug("failed to parse the token")
+			render.JSON(w, http.StatusUnauthorized, "failed to parse the token")
 			return
 		}